feat: added primary wallet screen

+++ app/src/main/java/com/degrin/bitcoinwallet/feature/wallet/domain/repository/WalletRepository.kt
@@ -1,5 +1,7 @@
 package com.degrin.bitcoinwallet.feature.wallet.domain.repository
 
+import java.math.BigDecimal
+
 interface WalletRepository {
-    suspend fun getBalance(): Long
+    suspend fun getBalance(): BigDecimal
 }
